+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/channel/list/viewmodel/ChannelListViewModel.kt
@@ -11,6 +11,8 @@ import io.getstream.chat.android.client.api.models.FilterObject
 import io.getstream.chat.android.client.api.models.QuerySort
 import io.getstream.chat.android.client.call.enqueue
 import io.getstream.chat.android.client.errors.ChatError
+import io.getstream.chat.android.client.logger.ChatLogger
+import io.getstream.chat.android.client.logger.TaggedLogger
 import io.getstream.chat.android.client.models.Channel
 import io.getstream.chat.android.client.models.Filters
 import io.getstream.chat.android.client.models.TypingEvent
@@ -52,6 +54,8 @@ public class ChannelListViewModel(
     public val paginationState: LiveData<PaginationState> = Transformations.distinctUntilChanged(paginationStateMerger)
     private val _errorEvents: MutableLiveData<Event<ErrorEvent>> = MutableLiveData()
     public val errorEvents: LiveData<Event<ErrorEvent>> = _errorEvents
+
+    private val logger: TaggedLogger = ChatLogger.get("ChannelListViewModel")
 
     private val filterLiveData: LiveData<FilterObject?> =
         filter?.let(::MutableLiveData) ?: chatDomain.user.map(Filters::defaultChannelListFilter).asLiveData()
@@ -97,6 +101,8 @@ public class ChannelListViewModel(
                 paginationStateMerger.addSource(queryChannelsController.endOfChannels.asLiveData()) { endOfChannels ->
                     setPaginationState { copy(endOfChannels = endOfChannels) }
                 }
+            } else {
+                logger.logE("Could not query channels. Error: ${queryChannelsControllerResult.error()}")
             }
         }
     }
@@ -129,28 +135,47 @@ public class ChannelListViewModel(
 
     public fun leaveChannel(channel: Channel) {
         chatDomain.leaveChannel(channel.cid).enqueue(
-            onError = { _errorEvents.postValue(Event(ErrorEvent.LeaveChannelError(it))) }
+            onError = { chatError ->
+                logger.logE("Could not leave channel with id: ${channel.id}. Error: ${chatError.message}. Cause: ${chatError.cause?.message}")
+                _errorEvents.postValue(Event(ErrorEvent.LeaveChannelError(chatError)))
+            }
         )
     }
 
     public fun deleteChannel(channel: Channel) {
         chatDomain.deleteChannel(channel.cid).enqueue(
-            onError = { _errorEvents.postValue(Event(ErrorEvent.DeleteChannelError(it))) }
+            onError = { chatError ->
+                logger.logE("Could not delete channel with id: ${channel.id}. Error: ${chatError.message}. Cause: ${chatError.cause?.message}")
+                _errorEvents.postValue(Event(ErrorEvent.DeleteChannelError(chatError)))
+            }
         )
     }
 
     public fun hideChannel(channel: Channel) {
         chatDomain.hideChannel(channel.cid, true).enqueue(
-            onError = { _errorEvents.postValue(Event(ErrorEvent.HideChannelError(it))) }
+            onError = { chatError ->
+                logger.logE("Could not hide channel with id: ${channel.id}. Error: ${chatError.message}. Cause: ${chatError.cause?.message}")
+                _errorEvents.postValue(Event(ErrorEvent.HideChannelError(chatError)))
+            }
         )
     }
 
     public fun markAllRead() {
-        chatDomain.markAllRead().enqueue()
+        chatDomain.markAllRead().enqueue(
+            onError = { chatError ->
+                logger.logE("Could not mark all messages as read. Error: ${chatError.message}. Cause: ${chatError.cause?.message}")
+            }
+        )
     }
 
     private fun requestMoreChannels() {
-        filterLiveData.value?.let { chatDomain.queryChannelsLoadMore(it, sort).enqueue() }
+        filterLiveData.value?.let { filter ->
+            chatDomain.queryChannelsLoadMore(filter, sort).enqueue(
+                onError = { chatError ->
+                    logger.logE("Could not load more channels. Error: ${chatError.message}. Cause: ${chatError.cause?.message}")
+                }
+            )
+        }
     }
 
     private fun setPaginationState(reducer: PaginationState.() -> PaginationState) {